+++ src/test/java/mafia/mafiatogether/room/repository/RoomManagerTest.java
@@ -1,7 +1,6 @@
-package mafia.mafiatogether.domain;
+package mafia.mafiatogether.room.repository;
 
 import mafia.mafiatogether.room.domain.RoomInfo;
-import mafia.mafiatogether.room.repository.RoomManager;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
